fix more /Wall compiler warnings in "sys"

+++ modules/c++/sys/source/SemaphoreWin32.cpp
@@ -31,7 +31,7 @@ sys::SemaphoreWin32::SemaphoreWin32(unsigned int count, size_t _maxCount)
 {
     // Ensure maxCount never becomes negative due to casting between signed/unsigned types
     const LONG maxLong = std::numeric_limits<LONG>::max();
-    LONG maxCount = (_maxCount > maxLong) ? maxLong : static_cast<LONG>(_maxCount);
+    LONG maxCount = (_maxCount > static_cast<size_t>(maxLong) ? maxLong : static_cast<LONG>(_maxCount);
 
     mNative = CreateSemaphoreA(NULL, static_cast<LONG>(count), maxCount, NULL);
     if (mNative == NULL)
